+++ packages/cryptoplease/lib/features/activities/src/activity.dart
@@ -3,24 +3,24 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import '../../outgoing_direct_payments/module.dart';
 import '../../outgoing_split_key_payments/module.dart';
 
-part 'pending_activity.freezed.dart';
+part 'activity.freezed.dart';
 
 @freezed
-class PendingActivity with _$PendingActivity {
-  const factory PendingActivity.outgoingPaymentRequest({
+class Activity with _$Activity {
+  const factory Activity.outgoingPaymentRequest({
     required String id,
     required DateTime created,
-  }) = OPRPendingActivity;
+  }) = OPRActivity;
 
-  const factory PendingActivity.outgoingDirectPayment({
+  const factory Activity.outgoingDirectPayment({
     required String id,
     required DateTime created,
     required OutgoingDirectPayment data,
-  }) = ODPPendingActivity;
+  }) = ODPActivity;
 
-  const factory PendingActivity.outgoingSplitKeyPayment({
+  const factory Activity.outgoingSplitKeyPayment({
     required String id,
     required DateTime created,
     required OutgoingSplitKeyPayment data,
-  }) = OSKPPendingActivity;
+  }) = OSKPActivity;
 }