+++ packages/solana/test/metaplex/instructions/create_metadata_account_test.dart
@@ -12,49 +12,54 @@ Future<void> main() async {
   // TODO(KB): Setup localnet with the metaplex program deployed.
   final client = createTestSolanaClient(useLocal: false);
   final owner = await Ed25519HDKeyPair.random();
-  await client.requestAirdrop(
-    address: owner.publicKey,
-    lamports: 1 * lamportsPerSol,
-    commitment: Commitment.confirmed,
+
+  test(
+    'creates metadata account',
+    () async {
+      await client.requestAirdrop(
+        address: owner.publicKey,
+        lamports: 1 * lamportsPerSol,
+        commitment: Commitment.confirmed,
+      );
+      final mint = await client.initializeMint(
+        mintAuthority: owner,
+        decimals: 0,
+        commitment: Commitment.confirmed,
+      );
+
+      final data = CreateMetadataAccountV3Data(
+        name: name,
+        symbol: symbol,
+        uri: uri,
+        sellerFeeBasisPoints: 550,
+        isMutable: false,
+        colectionDetails: false,
+      );
+      final instruction = await createMetadataAccountV3(
+        mint: mint.address,
+        mintAuthority: owner.publicKey,
+        payer: owner.publicKey,
+        updateAuthority: owner.publicKey,
+        data: data,
+      );
+
+      final message = Message.only(instruction);
+
+      await client.sendAndConfirmTransaction(
+        message: message,
+        signers: [owner],
+        commitment: Commitment.confirmed,
+      );
+
+      final metadata = await client.rpcClient.getMetadata(
+        mint: mint.address,
+        commitment: Commitment.confirmed,
+      );
+
+      expect(metadata?.name, name);
+      expect(metadata?.symbol, symbol);
+      expect(metadata?.uri, uri);
+    },
+    skip: 'Setup localnet with the metaplex program deployed.',
   );
-  final mint = await client.initializeMint(
-    mintAuthority: owner,
-    decimals: 0,
-    commitment: Commitment.confirmed,
-  );
-
-  test('creates metadata account', () async {
-    final data = CreateMetadataAccountV3Data(
-      name: name,
-      symbol: symbol,
-      uri: uri,
-      sellerFeeBasisPoints: 550,
-      isMutable: false,
-      colectionDetails: false,
-    );
-    final instruction = await createMetadataAccountV3(
-      mint: mint.address,
-      mintAuthority: owner.publicKey,
-      payer: owner.publicKey,
-      updateAuthority: owner.publicKey,
-      data: data,
-    );
-
-    final message = Message.only(instruction);
-
-    await client.sendAndConfirmTransaction(
-      message: message,
-      signers: [owner],
-      commitment: Commitment.confirmed,
-    );
-
-    final metadata = await client.rpcClient.getMetadata(
-      mint: mint.address,
-      commitment: Commitment.confirmed,
-    );
-
-    expect(metadata?.name, name);
-    expect(metadata?.symbol, symbol);
-    expect(metadata?.uri, uri);
-  });
 }